+++ src/market_data/mod.rs
@@ -1,2 +1,3 @@
+pub mod publish;
 pub mod sidefx;
 pub mod track;
